refactor: change face auth 1 steps flow, PopUpButton
