+++ src/screens/FaceRecognitionFlow/FaceAuthUserConsent/index.tsx
@@ -4,6 +4,7 @@ import {COLORS} from '@constants/colors';
 import {commonStyles} from '@constants/styles';
 import {Header} from '@navigation/components/Header';
 import {CountrySelect} from '@screens/FaceRecognitionFlow/FaceAuthUserConsent/CountrySelect';
+import {SelfieGuidelines} from '@screens/FaceRecognitionFlow/FaceAuthUserConsent/SelfieGuidelines';
 import {UserConsent} from '@screens/FaceRecognitionFlow/FaceAuthUserConsent/UserConsent';
 import {t} from '@translations/i18n';
 import React, {useState} from 'react';
@@ -13,8 +14,10 @@ type Props = {
   updateKycStepPassed: () => void;
 };
 
+type Step = 'consent' | 'country' | 'guidelines';
+
 export function FaceAuthUserConsent({updateKycStepPassed}: Props) {
-  const [consentPassed, setConsentPassed] = useState(false);
+  const [currentStep, setCurrentStep] = useState<Step>('consent');
 
   return (
     <View style={commonStyles.flexOne}>
@@ -23,14 +26,14 @@ export function FaceAuthUserConsent({updateKycStepPassed}: Props) {
         title={t('face_auth.header')}
         backgroundColor={'transparent'}
       />
-      {consentPassed ? (
-        <CountrySelect onContinue={updateKycStepPassed} />
-      ) : (
-        <UserConsent
-          updateKycStepPassed={() => {
-            setConsentPassed(true);
-          }}
-        />
+      {currentStep === 'consent' && (
+        <UserConsent onStepComplete={() => setCurrentStep('country')} />
+      )}
+      {currentStep === 'country' && (
+        <CountrySelect onStepComplete={() => setCurrentStep('guidelines')} />
+      )}
+      {currentStep === 'guidelines' && (
+        <SelfieGuidelines onStepComplete={updateKycStepPassed} />
       )}
     </View>
   );
