+++ src/screens/SettingsFlow/NotificationSettings/components/NotificationControls/hooks/useConfirmNotificationsDlg.tsx
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: ice License 1.0
 
-import {MainStackParamList} from '@navigation/Main';
-import {useNavigation} from '@react-navigation/native';
-import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {
   DEFAULT_DIALOG_NO_BUTTON,
   DEFAULT_DIALOG_YES_BUTTON,
-} from '@screens/Modals/PopUp/components/PopUpButton';
+} from '@components/Buttons/PopUpButton';
+import {MainStackParamList} from '@navigation/Main';
+import {useNavigation} from '@react-navigation/native';
+import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {PermissionsActions} from '@store/modules/Permissions/actions';
 import {t} from '@translations/i18n';
 import {useDispatch} from 'react-redux';
